Rename AnalyzeResult to RiskProfile
