+++ src/main/java/no/oslomet/aaas/controller/AnalyzationController.java
@@ -1,6 +1,6 @@
 package no.oslomet.aaas.controller;
 
-import no.oslomet.aaas.model.AnalyzeResult;
+import no.oslomet.aaas.model.RiskProfile;
 import no.oslomet.aaas.model.Request;
 import no.oslomet.aaas.service.AnalyzationService;
 import no.oslomet.aaas.service.LoggerService;
@@ -29,7 +29,7 @@ public class AnalyzationController {
     }
 
     @PostMapping
-    public AnalyzeResult getPayloadAnalyze(@Valid @RequestBody Request payload, HttpServletRequest request) {
+    public RiskProfile getPayloadAnalyze(@Valid @RequestBody Request payload, HttpServletRequest request) {
         loggerService.loggPayload(payload, request.getRemoteAddr(), AnalyzationController.class);
         return analyzationService.analyze(payload);
     }
