+++ src/main/java/no/oslomet/aaas/service/AnonymizationService.java
@@ -20,15 +20,15 @@ public class AnonymizationService {
 
     public AnonymizationResultPayload anonymize(Request payload){
 
-       AnalyzeResult beforeAnalyzeResult = analyzer.analyze(payload);
+       RiskProfile beforeRiskProfile = analyzer.analyze(payload);
 
         AnonymizeResult result = this.anonymizer.anonymize(payload);
 
         Request afterAnalysisPayload =
                 new Request(result.getData(), payload.getAttributes(), null);
-        AnalyzeResult afterAnalyzeResult = analyzer.analyze(afterAnalysisPayload);
+        RiskProfile afterRiskProfile = analyzer.analyze(afterAnalysisPayload);
 
-        return new AnonymizationResultPayload(result, beforeAnalyzeResult.getReIdentificationRisk(), afterAnalyzeResult.getReIdentificationRisk());
+        return new AnonymizationResultPayload(result, beforeRiskProfile.getReIdentificationRisk(), afterRiskProfile.getReIdentificationRisk());
     }
 
 }
