+++ src/test/java/no/oslomet/aaas/controller/AnalyzationControllerTest.java
@@ -1,7 +1,7 @@
 package no.oslomet.aaas.controller;
 
 import no.oslomet.aaas.GenerateTestData;
-import no.oslomet.aaas.model.AnalyzeResult;
+import no.oslomet.aaas.model.RiskProfile;
 import no.oslomet.aaas.model.Request;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -33,7 +33,7 @@ class AnalyzationControllerTest {
     @Test
     void getPayloadAnalyze() {
 
-        ResponseEntity<AnalyzeResult> responseEntity = restTemplate.postForEntity("/api/analyze",testPayload, AnalyzeResult.class);
+        ResponseEntity<RiskProfile> responseEntity = restTemplate.postForEntity("/api/analyze",testPayload, RiskProfile.class);
         assertNotNull(responseEntity);
         assertSame(HttpStatus.OK , responseEntity.getStatusCode());
         var resultData = responseEntity.getBody();
